Squashed commit of the following:
commit d43b2707ff563f08d5102d011cc40f8dfce60be3
Author: 1vals <1valsworld@gmail.com>
Date:   Tue Sep 8 22:35:56 2026 -0400

    refactor processor state implementation

commit 870ab8b7b6015be8a565cf352ed07a3345c1a1b6
Author: 1vals <1valsworld@gmail.com>
Date:   Tue Sep 8 22:11:47 2026 -0400

    create bandState struct

    contains 2 compressor states & a frequency float

commit f4316c0a0a050731b0793805019325bb0a001216
Author: 1vals <1valsworld@gmail.com>
Date:   Tue Sep 8 22:09:54 2026 -0400

    create functions for param listeners in apvts wrapper

commit 546e62516d9e5cdb4524277e8b3095a84caff288
Author: 1vals <1valsworld@gmail.com>
Date:   Tue Sep 8 20:45:16 2026 -0400

    get strip to work

commit 2b08ff9c886104051f6f1ac3aa1dcceba715ba74
Author: 1vals <1valsworld@gmail.com>
Date:   Tue Sep 8 20:24:42 2026 -0400

    Update SecondOrderButterworth.h

commit af57f1f07b9b410e6195a03956fedfe89ef8f716
Author: 1vals <1valsworld@gmail.com>
Date:   Tue Sep 8 06:06:29 2026 -0400

    attempting to implement upwards comp in strip

    not working, i think either the formula is wrong or smth else

commit 853b715808323d30fe54e66d996277b094a6bdf6
Author: 1vals <1valsworld@gmail.com>
Date:   Mon Sep 7 10:23:08 2026 -0400

    implement compressor in strip

commit fda1d63408d4387727298704d16966d9627e2587
Merge: b6e7eb0 057b3f2
Author: 1vals <1valsworld@gmail.com>
Date:   Mon Sep 7 09:58:04 2026 -0400

    Merge branch 'feature/bandstrip' of https://github.com/1vals/vOTT into feature/bandstrip

commit b6e7eb0df3ff8aac415ec6f503b402989185ea52
Author: 1vals <1valsworld@gmail.com>
Date:   Mon Sep 7 09:57:39 2026 -0400

    update BandStrip()

commit ef1b564a6f86826e3a3b8b55d682f1644a4e245e
Author: 1vals <1valsworld@gmail.com>
Date:   Mon Sep 7 09:56:10 2026 -0400

    begin implementation of BandStrip

commit 222733c7bd02033c209b2b46e38ea03e5867d847
Author: 1vals <1valsworld@gmail.com>
Date:   Mon Sep 7 09:55:13 2026 -0400

    refactor updateParams()

commit a699dc3ae3c5590af4f5f46077ded1e0381cb477
Author: 1vals <1valsworld@gmail.com>
Date:   Mon Sep 7 09:54:15 2026 -0400

    Delete LinkwitzRileyCrossover.h

commit d9a1a2836570ede4fedfaa1f2c0be0408d691e33
Author: 1vals <1valsworld@gmail.com>
Date:   Sun Sep 6 15:32:52 2026 -0400

    Begin creating bandstrip

commit 057b3f22ff3c26b4026ef6cd164675b7eea1f78b
Author: 1vals <1valsworld@gmail.com>
Date:   Sun Sep 6 15:32:52 2026 -0400

    Begin creating bandstrip

+++ src/common/dsp/dynamics/Envelope.cpp
@@ -11,7 +11,7 @@ void Envelope::prepare(int sr) {
 float Envelope::calculateConstant(float ms) const noexcept {
     // would this run better async, or is that redundant?
     return ms < static_cast<float>(1.0e-3) ? 0
-                                           : static_cast<float>(std::exp(expFactor / ms));
+                                           : std::exp(expFactor / ms);
 }
 
 void Envelope::process(const juce::dsp::ProcessContextReplacing<float>& context) {
@@ -42,8 +42,7 @@ float Envelope::processSample(int channel, float inputValue) {
     // otherwise the release constant is applied
     float cte = (inputValue > y[(size_t) channel] ? constantAtk : constantRls);
 
-    // add the input value to the constant, then multiply it by the difference between the current
-    // and old values
+    // add the input value to the constant, then multiply it by the difference between the curren and old values
     float result = inputValue + cte * (y[(size_t) channel] - inputValue);
     y[(size_t) channel] = result;
